OI Convert CSV: Fix import of fractional quantities.
git-svn-id: https://openintents.googlecode.com/svn/trunk@3138 72b678ce-9140-0410-bee8-679b907dd61a

+++ convertcsv/ConvertCSV/src/org/openintents/convertcsv/shoppinglist/ImportCsv.java
@@ -142,11 +142,11 @@ public class ImportCsv {
 				}
 			}
 			
-			long quantity;
+			double quantity;
 			try {
-			    quantity = Integer.parseInt(nextLine[4]); // Quantity
+			    quantity = Double.parseDouble(nextLine[4]); // Quantity
 			} catch (java.lang.NumberFormatException nfe){
-				quantity = 1;		
+				quantity = 1.0;		
 			}
 			long priority;
 			try {
